port code for reading the SD card from LdB's SD_FAT32 implementation from https://github.com/LdB-ECM/Raspberry-Pi/blob/33b176202b5ee12e19ed8757526190e62efb2975/SD_FAT32/SDCard.c also some other changes to implement what the driver needs

+++ src/rpi-log.c
@@ -43,11 +43,11 @@ void RPI_vLog(const char* source, unsigned level, const char* message, va_list v
 	RPI_TermSetBackgroundColor(old_bg);
 }
 
-void RPI_LogDump(const char* source, const int* buffer, unsigned length) {
+void RPI_LogDump(const char* source, const uint8_t* buffer, unsigned length) {
 	RPI_LogDumpColumns(source, buffer, length, 0);
 }
 
-void RPI_LogDumpColumns(const char* source, const int* buffer, unsigned length, unsigned columns) {
+void RPI_LogDumpColumns(const char* source, const uint8_t* buffer, unsigned length, unsigned columns) {
 	int old_fg = RPI_TermGetTextColor();
 	int old_bg = RPI_TermGetBackgroundColor();
 	RPI_TermSetBackgroundColor(COLORS_BLACK);
@@ -55,7 +55,7 @@ void RPI_LogDumpColumns(const char* source, const int* buffer, unsigned length, 
 	printf("[%s]: Dumping %u bytes at 0x%0X:\n", source, length, buffer);
 
 	while(length-- > 0) {
-		printf("%02X ", *buffer++);
+		printf("%02X ", (uint8_t)*buffer++);
 		if(columns != 0 && length % columns == 0) printf("\n");
 	}
